fix(sql-gateway): 跨源归并列对齐与类型转换修复
- 新增 ColumnAligner：按列名对齐结果集，支持大小写无关匹配与并集列计算
- 新增 ColumnTypeConverter：统一类型转换（数值→BigDecimal，字符串trim，时间→ISO字符串）
- CrossSourceUnionEngine：删除 validateColumnCounts，改用 alignAndConvertAll 按列名对齐+类型归一化
- CrossSourceJoinEngine：combineColumns 大小写无关冲突检测，combineRow 类型归一化
- CrossSourceJoinEngine：normalizeKey/compareValues/compareForSort 统一经 ColumnTypeConverter 归一化
- CrossSourceJoinEngine：Hash/Merge Join 跳过 null 键（SQL 语义 NULL=NULL 为 UNKNOWN）
- 更新 CrossSourceJoinEngineTest/CrossSourceExecutorTest 断言为 BigDecimal
- 新增 ColumnAlignerTest/ColumnTypeConverterTest/CrossSourceMergeAlignmentTest 共 63 个测试
- ROADMAP.md 标记任务完成

+++ platform/sql-gateway/src/test/java/com/levango7/dataenginebdp/sqlgateway/crosssource/CrossSourceExecutorTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -63,10 +64,12 @@ class CrossSourceExecutorTest {
 
         assertThat(result.getRowCount()).isEqualTo(4);
         assertThat(result.getSource()).isEqualTo("merged");
-        // 验证所有 id 都出现
+        // 验证所有 id 都出现（数值经类型归一化为 BigDecimal）
         List<Object> ids = new ArrayList<>();
         result.getRows().forEach(row -> ids.add(row.get(0)));
-        assertThat(ids).containsExactlyInAnyOrder(1, 2, 3, 4);
+        assertThat(ids).containsExactlyInAnyOrder(
+                BigDecimal.valueOf(1), BigDecimal.valueOf(2),
+                BigDecimal.valueOf(3), BigDecimal.valueOf(4));
     }
 
     @Test
